tools/bm-maintenance: only use one host mount

+++ packages/cleanup-bare-metal.sh
@@ -39,7 +39,7 @@ mapfile -t unusedRuntimeClasses < <(
         tr ' ' '\n' |
         grep '^contrast-cc' || true
       # Get all (maybe already deleted) runtime classes with a reference in /opt/edgeless
-      ls -1 "${OPTEDGELESS}"
+      ls -1 "${HOST_MOUNT:?}/opt/edgeless"
     } | sort -u
   )
 )
@@ -50,23 +50,23 @@ for runtimeClass in "${unusedRuntimeClasses[@]}"; do
   kubectl delete runtimeclass "${runtimeClass}" || true
 
   # Delete unused files
-  if [[ -d "${OPTEDGELESS}/${runtimeClass}" ]]; then
-    echo "Deleting files from ${OPTEDGELESS}/${runtimeClass} ..."
-    rm -rf "${OPTEDGELESS:?}/${runtimeClass}"
+  if [[ -d "${HOST_MOUNT}/opt/edgeless/${runtimeClass}" ]]; then
+    echo "Deleting files from ${HOST_MOUNT}/opt/edgeless/${runtimeClass} ..."
+    rm -rf "${HOST_MOUNT:?}/opt/edgeless/${runtimeClass}"
   fi
-  if [[ -d "/var/lib/${SNAPSHOTTER}-snapshotter" ]]; then
-    echo "Deleting files from /var/lib/${SNAPSHOTTER}-snapshotter/${runtimeClass} ..."
-    rm -rf "/var/lib/${SNAPSHOTTER}-snapshotter/${runtimeClass}"
+  if [[ -d "${HOST_MOUNT}/var/lib/${SNAPSHOTTER}-snapshotter" ]]; then
+    echo "Deleting files from ${HOST_MOUNT}/var/lib/${SNAPSHOTTER}-snapshotter/${runtimeClass} ..."
+    rm -rf "${HOST_MOUNT:?}/var/lib/${SNAPSHOTTER}-snapshotter/${runtimeClass}"
   fi
 
   # Remove references from containerd config
-  echo "Removing ${runtimeClass} from ${CONFIG} ..."
+  echo "Removing ${runtimeClass} from ${HOST_MOUNT}/${CONFIG} ..."
   # First try config v3 path. If this fails, try config v2 path.
-  dasel delete --file "${CONFIG}" --indent 0 --read toml --write toml "plugins.io\.containerd\.cri\.v1\.runtime.containerd.runtimes.${runtimeClass}" ||
-    dasel delete --file "${CONFIG}" --indent 0 --read toml --write toml "plugins.io\.containerd\.grpc\.v1\.cri.containerd.runtimes.${runtimeClass}" ||
+  dasel delete --file "${HOST_MOUNT:?}/${CONFIG}" --indent 0 --read toml --write toml "plugins.io\.containerd\.cri\.v1\.runtime.containerd.runtimes.${runtimeClass}" ||
+    dasel delete --file "${HOST_MOUNT:?}/${CONFIG}" --indent 0 --read toml --write toml "plugins.io\.containerd\.grpc\.v1\.cri.containerd.runtimes.${runtimeClass}" ||
     true
 
-  dasel delete --file "${CONFIG}" --indent 0 --read toml --write toml "proxy_plugins.${SNAPSHOTTER}-${runtimeClass}" 2>/dev/null || true
+  dasel delete --file "${HOST_MOUNT:?}/${CONFIG}" --indent 0 --read toml --write toml "proxy_plugins.${SNAPSHOTTER}-${runtimeClass}" 2>/dev/null || true
 done
 
 echo "Cleanup finished"
